requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
google-api-python-client>=2.108.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
python-dotenv>=1.0.0
pydantic[email]>=2.5.0
python-multipart>=0.0.6
email-validator>=2.0.0
pytz>=2024.1